Cambios Mayores

+++ metas/models/meta.js
@@ -1,7 +1,7 @@
 module.exports = function(sequelize, DataTypes) {
 	var Meta = sequelize.define('metas', {
 		descripcion: DataTypes.STRING,
-		valor_esperado: DataTypes.INTEGER,
+		valor_esperado: DataTypes.DOUBLE,
 		fecha_limite: DataTypes.DATE,
 		producto_id: DataTypes.INTEGER
 	});
